+++ changelog.md
@@ -1,2 +1,18 @@
-2.0.2
-  
+# Changelog
+
+## [2.0.3] - 2020-06-09
+
+### Added
+- Add function `util->getUserPhoto()`
+- Add phpUnit Test
+### Fixed
+
+- Rename class `utilities` on `modUtilities`
+- Rename class `utilitiesCsv` on `modUtilitiesCsv`
+- Fixed an error when duplicating a class
+
+## [2.0.2] - 2020-06-08
+
+### Changed
+
+- class csv add method `csv->toHtml('classname')`